isolated serialDriver
